+++ src/tedi/components/navigation/pagination/index.ts
@@ -1,2 +1,3 @@
 export * from './pagination';
+export * from './pagination-mobile-modal/pagination-mobile-modal';
 export * from './use-pagination';