+++ src/tedi/index.ts
@@ -31,6 +31,7 @@ export * from './components/notifications/toast/toast';
 export * from './components/cards/card';
 export * from './components/navigation/breadcrumbs';
 export * from './components/navigation/hash-trigger/hash-trigger';
+export * from './components/navigation/horizontal-stepper';
 export * from './components/navigation/link/link';
 export * from './components/navigation/pagination';
 export * from './components/navigation/tabs';
@@ -67,6 +68,7 @@ export * from './components/misc/affix/affix';
 export * from './components/misc/attachment';
 export * from './components/misc/ellipsis/ellipsis';
 export * from './components/layout/sidenav';
+export * from './components/layout/footer';
 
 export * from './helpers';